refactor: extract tile slice dispatch from impl.py into tile_slicing.py
Move the tile-specific subscript interception logic (try_tile_ref,
try_tile_slice) out of impl.py into a new simt/tile_slicing.py module.
Reduces impl.py complexity and keeps tile concerns co-located.
